Replaced pytz with zoneinfo in functions.py
From what I can tell online, pytz has been depreciated for a few years now and is not avaliable by default in the newest versions of Python. This change replaces it with zoneinfo, which should be compatable with all versions of Python3. From my testing it works exactly the same as it did previously, just one line shorter and with better compatability moving forward. I'm no Python expert, so if there's some problem with my changes or reason why keeping pytz is preferable feel free to disregard.